9272/move properties to service-api and change jwt properties usage
