+++ service-api/src/main/java/greencity/properties/GoogleProperties.java
