+++ service-api/src/main/java/greencity/properties/RemoteWebClientProperties.java
