+++ service-api/src/main/java/greencity/properties/TelegramProperties.java
