+++ service-api/src/main/java/greencity/properties/WayForPayProperties.java
